% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antpost.R
\name{antpost}
\alias{antpost}
\title{Digitize the left and right margins of a silhouette.}
\usage{
antpost(Dat, Add.Tip = T, tip.radius = 1e-04, Origin = c("Bottom", "Top"))
}
\arguments{
\item{Dat}{Black and white image in the form of a matrix filled with 0s and 1s.}

\item{Add.Tip}{Logical. Whether to add a tip to the ends of the silhouette.}

\item{tip.radius}{Radius of the tip added if Add.Tip = TRUE.}

\item{Origin}{Whether the top of the bottom is considered the origin.}
}
\value{
Five-column matrix containing y coordinates, silhouette left x coordinates, silhouette right x coordinates, silhouette center x coordinates, and the radius of the silhoutte at a given y.
}
\description{
Digitize the left and right margins of a silhouette.
}
\examples{
Used by paleomass()
}
