% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locfit_interpolate.R
\name{locfit_interpolate}
\alias{locfit_interpolate}
\title{Smooth the outline coordinates from topbottom() or antpost() using local regression.}
\usage{
locfit_interpolate(
  Mx,
  Nseg = c("asis", "round", "provided"),
  n.out = 1000,
  Use.Weight = F,
  nn = 0.1,
  h = 0
)
}
\arguments{
\item{Mx}{Output matrix from topbottom() or antpost().}

\item{Nseg}{How to deal with the number of segments}

\item{n.out}{Number of segments if Nseg = "provided"}

\item{Use.Weight}{Whether to add weights to the tips}

\item{nn}{Nearest neighbor component of locfit()}

\item{h}{Constant component of locfit()}
}
\value{
Five-column matrix as in the input matrix but with smoothed outcome
}
\description{
Smooth the outline coordinates from topbottom() or antpost() using local regression.
}
\examples{
Used by paleomass()
}
