% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh_body.R
\name{mesh_body}
\alias{mesh_body}
\title{Build a mesh3d object of the main body from two orthogonal silhouettes and superellipticla cross-sections}
\usage{
mesh_body(
  dat.lat,
  dat.dv,
  nb = 2,
  X.center = 0,
  Y.center = 0,
  Z.onset = 0,
  a.per.slice = 181,
  Mesh.Col = "royalblue",
  tol = 1e-04
)
}
\arguments{
\item{dat.lat}{Lateral view silhouette processed by topbottom() or locfit_interpolate().}

\item{dat.dv}{Ventral view silhouette processed by topbottom() or locfit_interpolate().}

\item{nb}{Exponent of the superellipse to be used. nb=2 makes it a normal ellipse.}

\item{X.center}{Center of global X coordinate to anchor the mesh.}

\item{Y.center}{Center of global Y coordinate to anchor the mesh.}

\item{Z.onset}{Onset of global Z coordinate to anchor the mesh.}

\item{a.per.slice}{Number of angular segments in a cross-section.}

\item{Mesh.Col}{Color of the resulting mesh3d object.}

\item{tol}{Tolerance limit used by vcgClean().}
}
\value{
A mesh3d model of the main body.
}
\description{
Build a mesh3d object of the main body from two orthogonal silhouettes and superellipticla cross-sections
}
\examples{
Used by paleomass()
}
