% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh_foil.R
\name{mesh_foil}
\alias{mesh_foil}
\title{Build a mesh3d model of fins using a planar silhouette and NACA 4-digit foil sections}
\usage{
mesh_foil(
  dat.fin,
  tf = 20,
  X.onset = 0,
  Y.onset = 0,
  Z.onset = 0,
  tol = 1e-04,
  Center = T,
  Thickest = 0.5,
  a.per.slice = 181,
  Mesh.Col = "royalblue",
  Cephalofoil = F
)
}
\arguments{
\item{dat.fin}{Planar view silhouette of the fin processed by antpost() or locfit_interpolate().}

\item{tf}{Maximum thickness of the foil section relative to chord in percentage, e.g., tf=20 would result in NACA 0020 foil sections.}

\item{X.onset}{Global X coordinate of the fin onset.}

\item{Y.onset}{Global Y coordinate of the fin onset.}

\item{Z.onset}{Global Z coordinate of the fin onset.}

\item{tol}{Tolerance limit used by vcgClean().}

\item{Center}{Whether to Center the fin along Y axis or not.}

\item{Thickest}{Value between 0 and 1 indicating the thickest part of the fin along its span. 0 for bottom and 1 for top.}

\item{a.per.slice}{Number of angular segments in a cross-section.}

\item{Mesh.Col}{Mesh.Col Color of the resulting mesh3d object.}

\item{Cephalofoil}{Logical. If the mesh is for a cephalofoil or not.}
}
\value{
A mesh3d object representing the fin.
}
\description{
Build a mesh3d model of fins using a planar silhouette and NACA 4-digit foil sections
}
\examples{
Used by paleomass()
}
