% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paleomass.R
\name{paleomass}
\alias{paleomass}
\title{Main function to build two 3D models from silhouettes of a marine vertebrate in a trial to bracket the body volume and surface area between the two.}
\usage{
paleomass(
  Folder = "Stenopterygius",
  File.Ext = c(".png", ".jpg", ".jpeg", ".bmp", ""),
  BodyL.File = "BodyL",
  BodyV.File = "BodyV",
  Cfin.File = "Cfin",
  Dfin.File = "Dfin",
  D2fin.File = "D2fin",
  Ffin.File = "Ffin",
  Hfin.File = "Hfin",
  Afin.File = "Afin",
  Ceph.File = "Ceph",
  n1 = 2,
  n2 = 3,
  fork.l = 1.19,
  dens = 0.95,
  Save.Csv = T,
  Interpolate = T,
  nn.b = 0.1,
  nn.f = 0.6,
  Use.Weight = F,
  Add.Tip = T,
  tip.radius = 1e-04,
  a.per.slice = 181,
  Plot.Result.Mesh = T,
  Save.Part.Mesh = F,
  Save.Total.Mesh = T,
  N.Mesh.Window = c("two", "one"),
  Mesh.Col = "royalblue",
  Body.Y.Centering = T,
  N.Body.Slice = c("round", "asis", "provided"),
  n.body.slice = 1000,
  N.Fin.Slice = c("asis", "round", "provided"),
  n.fin.slice = 500,
  cfin.adj.up = -5,
  cfin.onset = 870,
  cfin.rot = 0,
  cfin.sweep = 0,
  cfin.thick = 20,
  dfin.adj.up = 175,
  dfin.onset = 440,
  dfin.sweep = 0,
  dfin.thick = 20,
  d2fin.adj.up = 165,
  d2fin.onset = 600,
  d2fin.sweep = 0,
  d2fin.thick = 10,
  ffin.spread = pi/4,
  ffin.sweep = pi/6,
  ffin.adj.med = 30,
  ffin.adj.up = -67,
  ffin.onset = 300,
  ffin.thick = 20,
  hfin.spread = pi/6,
  hfin.sweep = pi/7,
  hfin.adj.med = 30,
  hfin.adj.up = -27,
  hfin.onset = 600,
  hfin.thick = 20,
  afin.adj.up = -25,
  afin.onset = 650,
  afin.sweep = 0,
  afin.thick = 10,
  ceph.adj.up = -5,
  ceph.onset = 870,
  ceph.rot = 0,
  ceph.sweep = 0,
  ceph.thick = 30,
  Test.Run = F
)
}
\arguments{
\item{Folder}{Name of the sub-directory of the working directory storing image files.}

\item{File.Ext}{File extension of the image file. Either ".bmp",".jpg",".png", or "", the last one for without extensions.}

\item{BodyL.File}{Name of the image file for the lateral silhouette of the body, without extension.}

\item{BodyV.File}{Name of the image file for the ventral silhouette of the body, without extension.}

\item{Cfin.File}{Name of the image file for the planar silhouette of the caudal fin, without extension.}

\item{Dfin.File}{. Name of the image file for the planar silhouette of the dorsal fin, if any, without extension.}

\item{D2fin.File}{Name of the image file for the planar silhouette of the second dorsal fin, if any, without extension.}

\item{Ffin.File}{Name of the image file for the planar silhouette of the forefin/pectoral fin/pectoral flipper.}

\item{Hfin.File}{Name of the image file for the planar silhouette of the hindfin/pelvic fin/pelvic flipper, without extension.}

\item{Afin.File}{Name of the image file for the planar silhouette of the anal fin, if any, without extension.}

\item{Ceph.File}{Name of the image file for the planar silhouette of the cephalofoil, if any, without extension.}

\item{n1}{Superelliptical exponent used in the first model. A value of 2.0 is recommended for marine tetrapods.}

\item{n2}{Superelliptical exponent used in the second model. A value of 3.0 is recommended for marine tetrapods.}

\item{fork.l}{Fork length of the marine vertebrate in meters.}

\item{dens}{Arbitrary specific gravity of whole body in g/cm^3.}

\item{Save.Csv}{Logical. Whether to save a CSV file containing measurement results.}

\item{Interpolate}{Logical. Whether to smooth the outline coordinates with local regression. TRUE is recommended.}

\item{nn.b}{Nearest neighbor component of locfit() used for body reconstruction when interpolating.}

\item{nn.f}{Nearest neighbor component of locfit() used for fin/flipper reconstruction when interpolating.}

\item{Use.Weight}{Logical. Whether to add weight to the tips when using locfit()}

\item{Add.Tip}{Logical. Whether to add tips and the two ends of an object. TRUE is recommended to close the holes generated at the tips. If TRUE, then a small circle of a radius specified by tip.radius (0.0001 pixel by default) is added, while the existing points are shifted inward by tip.radius. This affects the shape minimally while closing the tips.}

\item{tip.radius}{Radius of the tip to be added. A value of 0.0001 is recommended.}

\item{a.per.slice}{Number of angular division per superelliptical slice plus 1, that is, how many points are placed around a superellipse with one pair of redundant points. The default value is 181, i.e., there is a point every 2 degrees.}

\item{Plot.Result.Mesh}{Logical. Whether to plot resulting meshes.}

\item{Save.Part.Mesh}{Logical. Whether to save .ply files of each body part}

\item{Save.Total.Mesh}{Logical. Whether to save .ply files of combined body parts.}

\item{N.Mesh.Window}{A string of "two" or "one". Whether to plot two models side by side in one window or not. One window would often kill R.}

\item{Mesh.Col}{Color of the mesh in rgl window. Not reflected in the saved .ply files.}

\item{Body.Y.Centering}{Logical. Whether to center the body along the Y axis.}

\item{N.Body.Slice}{How to set the number of longitudinal segments. Possible values are: "asis"#’ @param #’ @param use the pixel count from input image; "round"(default)#’ @param #’ @param round the value to the closest 100; " provided"#’ @param #’ @param use the value provided in n.body.slice}

\item{n.body.slice}{Number of segments along the body axis. Moot unless N.Body.Slice = "provided".}

\item{N.Fin.Slice}{How to set the number of spanwise segments. Possible values are: "asis"#’ @param #’ @param use the pixel count from input image; "round"(default)#’ @param #’ @param round the value to the closest 100; " provided"#’ @param #’ @param use the value provided in n.fin.slice}

\item{n.fin.slice}{umber of segments along the span of the fin. Moot unless N.Fin.Slice = "provided".}

\item{cfin.adj.up}{Position adjustment along the Y axis (dorso-ventral axis) for the caudal fin, in pixels.}

\item{cfin.onset}{Position adjustment along the Z axis ( cranio-caudal axis) for the caudal fin, in pixels.}

\item{cfin.rot}{Rotation angle of the caudal fin around Z axis ( cranio-caudal axis). 0 for #’ @param vertical, pi/2 for horizontal fins.}

\item{cfin.sweep}{Sweep angle of the caudal fin around the X axis (bilateral axis).}

\item{cfin.thick}{Maximum thickness of the caudal fin relative to the chord in percentage.}

\item{dfin.adj.up}{Position adjustment along the Y axis (dorso-ventral axis) for the dorsal fin, in pixels.}

\item{dfin.onset}{Position adjustment along the Z axis ( cranio-caudal axis) for the dorsal fin, in pixels.}

\item{dfin.sweep}{Sweep angle of the dorsal fin around the X axis (bilateral axis).}

\item{dfin.thick}{Maximum thickness of the dorsal fin relative to the chord in percentage.}

\item{d2fin.adj.up}{Position adjustment along the Y (dorso-ventral axis) axis for the second dorsal fin, in pixels.}

\item{d2fin.onset}{Position adjustment along the Z axis ( cranio-caudal axis) for the secod dorsal fin, in pixels.}

\item{d2fin.sweep}{Sweep angle of the second dorsal fin around the X axis (bilateral axis).}

\item{d2fin.thick}{Maximum thickness of the second dorsal fin relative to the chord in percentage.}

\item{ffin.spread}{Rotation angle of the forefin around the Z axis ( cranio-caudal axis).}

\item{ffin.sweep}{Sweep angle of the forefin around the X axis (bilateral axis).}

\item{ffin.adj.med}{Position adjustment along the X axis (bilateral axis) for the forefin, in pixels.}

\item{ffin.adj.up}{Position adjustment along the Y axis (dorso-ventral axis) for the forefin, in pixels.}

\item{ffin.onset}{Position adjustment along the Z axis ( cranio-caudal axis) for the forefin, in pixels.}

\item{ffin.thick}{Maximum thickness of the forefin relative to the chord in percentage.}

\item{hfin.spread}{Rotation angle of the hindfin around the Z axis ( cranio-caudal axis).}

\item{hfin.sweep}{Sweep angle of the hindfin around the X axis (bilateral axis).}

\item{hfin.adj.med}{Position adjustment along the X axis (bilateral axis) for the hindfin, in pixels.}

\item{hfin.adj.up}{Position adjustment along the Y axis (dorso-ventral axis) for the hindfin, in pixels.}

\item{hfin.onset}{Position adjustment along the Z axis ( cranio-caudal axis) for the hindfin, in pixels.}

\item{hfin.thick}{Maximum thickness of the hindfin relative to the chord in percentage.}

\item{afin.adj.up}{Position adjustment along the Y axis (dorso-ventral axis) for the anal fin, in pixels.}

\item{afin.onset}{Position adjustment along the Z axis (cranio-caudal axis) for the anal fin, in pixels.}

\item{afin.sweep}{Sweep angle of the anal fin around the X axis (bilateral axis).}

\item{afin.thick}{Maximum thickness of the anal fin relative to the chord in percentage.}

\item{ceph.adj.up}{Position adjustment along the Y axis (dorso#-ventral axis) for the cephalofoil, in pixels.}

\item{ceph.onset}{Position adjustment along the Z axis (cranio-caudal axis) for the cephalofoil, in pixels.}

\item{ceph.rot}{Rotation angle of the cephalofoil around Z axis (cranio-caudal axis). pi/2 for horizontal cephalfoil.}

\item{ceph.sweep}{Sweep angle of the cephalofoil around the X axis (bilateral axis).}

\item{ceph.thick}{Maximum thickness of the cephalofoil relative to the chord in percentage.}

\item{Test.Run}{Whether to set the positions and rotations of fins for a test run}
}
\value{
A matrix containing measurement results.
}
\description{
Main function to build two 3D models from silhouettes of a marine vertebrate in a trial to bracket the body volume and surface area between the two.
}
\examples{
# Ichthyosaur Example
# Save the silhouette images for Stenopterygius in the current
# directory and make 3D models using them.
save_Stenopterygius()
paleomass()
#
# Plesiosaur Example
# Save the silhouette images for Plesiosaurus and make 3D models using them
save_Plesiosaurus()
paleomass(Folder="Plesiosaurus",fork.l=2.94,ffin.onset=450,
    hfin.onset=650,ffin.adj.med=10,hfin.adj.med=40,ffin.adj.up=-160,
    hfin.adj.up=-160,ffin.spread=pi/3,hfin.spread=pi/4)
#
# Hammerhead Example
# Save the silhouette images for Sphyrna and make 3D models using them
save_Sphryna()
paleomass(Folder="Sphyrna",fork.l=2,n2=1.8,
    ffin.onset=300, ffin.adj.up=-120,ffin.adj.med=40,ffin.spread=pi/4,
    ffin.sweep=-pi*0/180,
    hfin.onset=550,hfin.adj.up=10,hfin.adj.med=40,hfin.spread=pi/4,
    hfin.thick=10,hfin.sweep=pi*5/180,
    cfin.onset=920,cfin.adj.up=35,
    dfin.onset=350,dfin.adj.up=190,dfin.sweep=pi*5/180,
    d2fin.onset=800,d2fin.adj.up=200,,d2fin.sweep=pi*5/180,d2fin.thick=10,
    afin.onset=800,afin.adj.up=60,afin.sweep=pi*20/180,afin.thick=10,
    ceph.onset=-45,ceph.adj.up=120, ceph.rot=pi/2)
}
