% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plesiosaurus.R
\docType{data}
\name{Plesiosaurus}
\alias{Plesiosaurus}
\title{Image data set containing silhouettes of Plesiosaurus in cimg}
\format{
a data.frame containing four cimg objects:
\describe{
\item{BodyL}{Lateral silhouette of the main body}
\item{BodyV}{Ventral silhouette of the main body}
\item{Ffin}{Forefin/pectoral flipper planar silhouette}
\item{Hfin}{Hindfin/pelvic flipper planar silhouette}
}
}
\source{
{Motani, R. Swimming speed estimation of extinct marine reptiles: energetic approach revisited. Paleobiology 28, 251–262 (2006). https://doi.org/10.1666/0094-8373\%282002\%29028\%3C0251\%3ASSEOEM\%3E2.0.CO\%3B2}
}
\usage{
Plesiosaurus
}
\description{
Image data set containing silhouettes of Plesiosaurus in cimg
}
\examples{
data(Plesiosaurus)
}
\keyword{datasets}
