% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_Plesiopterys.R
\name{save_Plesiopterys}
\alias{save_Plesiopterys}
\title{Save the silhouette images for Plesiopterys in a subfolder of the current directory}
\usage{
save_Plesiopterys()
}
\value{
None
}
\description{
Save the silhouette images for Plesiopterys in a subfolder of the current directory
}
\examples{
save_Plesiopterys()
}
