% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_Plesiopteryx.R
\name{save_Plesiopteryx}
\alias{save_Plesiopteryx}
\title{Save the silhouette images for Plesiopteryx in a subfolder of the current directory}
\usage{
save_Plesiopteryx()
}
\value{
None
}
\description{
Save the silhouette images for Plesiopteryx in a subfolder of the current directory
}
\examples{
save_Plesiopteryx()
}
