% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_Plesiosaurus.R
\name{save_Plesiosaurus}
\alias{save_Plesiosaurus}
\title{Save the silhouette images for Plesiosaurus in a subfolder of the current directory}
\usage{
save_Plesiosaurus()
}
\value{
None
}
\description{
Save the silhouette images for Plesiosaurus in a subfolder of the current directory
}
\examples{
save_Plesiosaurus()
}
