% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_Sphyrna.R
\name{save_Sphyrna}
\alias{save_Sphyrna}
\title{Save the silhouette images for Sphyrna in a subfolder of the current directory}
\usage{
save_Sphyrna()
}
\value{
None
}
\description{
Save the silhouette images for Sphyrna in a subfolder of the current directory
}
\examples{
save_Sphyrna()
}
