% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_Stenopterygius.R
\name{save_Stenopterygius}
\alias{save_Stenopterygius}
\title{Save the silhouette images for Stenopterygius in a subfolder of the current directory}
\usage{
save_Stenopterygius()
}
\value{
None
}
\description{
Save the silhouette images for Stenopterygius in a subfolder of the current directory
}
\examples{
save_Stenopterygius()
}
