% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sphyrna.R
\docType{data}
\name{Sphyrna}
\alias{Sphyrna}
\title{Image data set containing silhouettes of Sphyrna tudes in cimg}
\format{
a data.frame containing nine cimg objects:
\describe{
\item{BodyL}{Lateral silhouette of the main body}
\item{BodyV}{Ventral silhouette of the main body}
\item{Cfin}{Caudal fin planar silhouette}
\item{Dfin}{Dorsal fin planar silhouette}
\item{D2fin}{Second dorsal fin planar silhouette}
\item{Ffin}{Forefin/pectoral flipper planar silhouette}
\item{Hfin}{Hindfin/pelvic flipper planar silhouette}
\item{Afin}{Anal fin planar silhouette}
\item{Ceph}{Cephalofoil planar silhouette}
}
}
\source{
{Created in-house for package deomonstration purposes}
}
\usage{
Sphyrna
}
\description{
Image data set containing silhouettes of Sphyrna tudes in cimg
}
\examples{
data(Sphyrna)
}
\keyword{datasets}
