% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topbottom.R
\name{topbottom}
\alias{topbottom}
\title{Digitize the top and bottom outlines of a silhouette.}
\usage{
topbottom(Dat, Add.Tip = T, tip.radius = 1e-04, Origin = c("Bottom", "Top"))
}
\arguments{
\item{Dat}{Black and white image in the form of a matrix filled with 0s and 1s.}

\item{Add.Tip}{Logical. Whether to add a tip to the ends of the silhouette.}

\item{tip.radius}{Radius of the tip added if Add.Tip = TRUE.}

\item{Origin}{Whether the top of the bottom is considered the origin.}
}
\value{
Five-column matrix containing x coordinates, silhouette bottom y coordinates, silhouette top y coordinates, silhouette center y coordinates, and the radius of the silhoutte at a given x.
}
\description{
Digitize the top and bottom outlines of a silhouette.
}
\examples{
Used by paleomass()
}
