% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paleomass.R
\name{paleomass}
\alias{paleomass}
\title{Main function to build two 3D models from silhouettes of a marine vertebrate in a trial to bracket the body volume and surface area between the two.}
\usage{
paleomass(
  Folder = "Stenopterygius",
  File.Ext = c(".png", ".jpg", ".jpeg", ".bmp", ""),
  BodyL.File = "BodyL",
  BodyV.File = "BodyV",
  Cfin.File = "Cfin",
  Dfin.File = "Dfin",
  D2fin.File = "D2fin",
  Ffin.File = "Ffin",
  Hfin.File = "Hfin",
  Afin.File = "Afin",
  Ceph.File = "Ceph",
  Save.Csv = T,
  Save.Part.Mesh = F,
  Save.Total.Mesh = T,
  Plot.Result.Mesh = T,
  N.Mesh.Window = c("two", "one"),
  Mesh.Col = "royalblue",
  n1 = 2,
  n2 = 3,
  body.axis.l = 1.19,
  dens = 1.027,
  Interpolate = T,
  nn.b = 0.1,
  nn.f = 0.1,
  Use.Weight = F,
  Add.Tip = T,
  tip.radius = 1e-04,
  a.per.slice = 181,
  Body.Y.Centering = T,
  N.Body.Slice = c("round", "asis", "provided"),
  n.body.slice = 3000,
  N.Fin.Slice = c("asis", "round", "provided"),
  n.fin.slice = 500,
  cfin.adj.lat = 0,
  cfin.adj.up = 10,
  cfin.onset = 0.87,
  cfin.roll = 0,
  cfin.pitch = 0,
  cfin.yaw = 0,
  cfin.thick = 20,
  dfin.adj.up = -45,
  dfin.onset = 0.501,
  dfin.roll = 0,
  dfin.pitch = 0,
  dfin.yaw = 0,
  dfin.thick = 20,
  d2fin.adj.up = 0,
  d2fin.onset = 0.6,
  d2fin.roll = 0,
  d2fin.pitch = 0,
  d2fin.yaw = 0,
  d2fin.thick = 10,
  ffin.adj.lat = -200,
  ffin.adj.up = 230,
  ffin.onset = 0.349,
  ffin.roll = pi/4,
  ffin.pitch = pi/9,
  ffin.thick = 20,
  ffin.yaw = 0,
  hfin.adj.lat = -150,
  hfin.adj.up = 0,
  hfin.onset = 0.653,
  hfin.roll = pi/6,
  hfin.pitch = pi/7,
  hfin.thick = 20,
  hfin.yaw = 0,
  afin.adj.up = 0,
  afin.onset = 0.65,
  afin.roll = 0,
  afin.pitch = 0,
  afin.thick = 10,
  afin.yaw = 0,
  ceph.adj.up = 0,
  ceph.onset = -0.05,
  ceph.roll = pi/2,
  ceph.pitch = 0,
  ceph.thick = 30,
  ceph.yaw = 0,
  Test.Run = F
)
}
\arguments{
\item{Folder}{Name of the sub-directory of the working directory storing image files.}

\item{File.Ext}{File extension of the image file. Either ".bmp",".jpg",".png", or "", the last one for without extensions.}

\item{BodyL.File}{Name of the image file for the lateral silhouette of the body, without extension.}

\item{BodyV.File}{Name of the image file for the ventral silhouette of the body, without extension.}

\item{Cfin.File}{Name of the image file for the planar silhouette of the caudal fin, without extension.}

\item{Dfin.File}{. Name of the image file for the planar silhouette of the dorsal fin, if any, without extension.}

\item{D2fin.File}{Name of the image file for the planar silhouette of the second dorsal fin, if any, without extension.}

\item{Hfin.File}{Name of the image file for the planar silhouette of the hindfin/pelvic fin/pelvic flipper, without extension.}

\item{Afin.File}{Name of the image file for the planar silhouette of the anal fin, if any, without extension.}

\item{Ceph.File}{Name of the image file for the planar silhouette of the cephalofoil, if any, without extension.}

\item{Save.Csv}{Logical. Whether to save a CSV file containing measurement results.}

\item{Save.Part.Mesh}{Logical. Whether to save .ply files of each body part}

\item{Save.Total.Mesh}{Logical. Whether to save .ply files of combined body parts.}

\item{Plot.Result.Mesh}{Logical. Whether to plot resulting meshes.}

\item{N.Mesh.Window}{A string of "two" or "one". Whether to plot two models side by side in one window or not. One window would often kill R.}

\item{Mesh.Col}{Color of the mesh in rgl window. Not reflected in the saved .ply files.}

\item{n1}{Superelliptical exponent used in the first model. A value of 2.0 is recommended for marine tetrapods.}

\item{n2}{Superelliptical exponent used in the second model. A value of 3.0 is recommended for marine tetrapods.}

\item{body.axis.l}{Body axis length in the BodyL image in meters.}

\item{dens}{Arbitrary specific gravity of whole body in g/cm^3.}

\item{Interpolate}{Logical. Whether to smooth the outline coordinates with local regression. TRUE is recommended.}

\item{nn.b}{Nearest neighbor component of locfit() used for body reconstruction when interpolating.}

\item{nn.f}{Nearest neighbor component of locfit() used for fin/flipper reconstruction when interpolating.}

\item{Use.Weight}{Logical. Whether to add weight to the tips when using locfit()}

\item{Add.Tip}{Logical. Whether to add tips and the two ends of an object. TRUE is recommended to close the holes generated at the tips. If TRUE, then a small circle of a radius specified by tip.radius (0.0001 pixel by default) is added, while the existing points are shifted inward by tip.radius. This affects the shape minimally while closing the tips.}

\item{tip.radius}{Radius of the tip to be added. A value of 0.0001 is recommended.}

\item{a.per.slice}{Number of angular division per superelliptical slice plus 1, that is, how many points are placed around a superellipse with one pair of redundant points. The default value is 181, i.e., there is a point every 2 degrees.}

\item{Body.Y.Centering}{Logical. Whether to center the body along the Y axis.}

\item{N.Body.Slice}{How to set the number of longitudinal segments. Possible values are: "asis" use the pixel count from input image; "round"(default) round the value to the closest 100; " provided" use the value provided in n.body.slice}

\item{n.body.slice}{Number of segments along the body axis. Moot unless N.Body.Slice = "provided".}

\item{N.Fin.Slice}{How to set the number of spanwise segments. Possible values are: "asis" use the pixel count from input image; "round"(default) round the value to the closest 100; " provided" use the value provided in n.fin.slice}

\item{n.fin.slice}{umber of segments along the span of the fin. Moot unless N.Fin.Slice = "provided".}

\item{cfin.adj.lat}{Position adjustment along the X axis (bilateral axis) for the caudal fin, in pixels.}

\item{cfin.adj.up}{Position adjustment along the Y axis (dorso-ventral axis) for the caudal fin, in pixels.}

\item{cfin.onset}{Position adjustment along the Z axis (cranio-caudal axis) for the caudal fin, in fraction, usually between 0 and 1.}

\item{cfin.roll}{Rotation angle of the caudal fin around Z axis (cranio-caudal axis). 0 for vertical, pi/2 for horizontal fins.}

\item{cfin.pitch}{Rotation angle of the caudal fin around the X axis (bilateral axis). Becomes a yaw angle when cfin.roll = pi/2.}

\item{cfin.yaw}{Yaw angle of caudal fin around the Y axis (dorso-ventral axis).}

\item{cfin.thick}{Maximum thickness of the caudal fin relative to the chord in percentage.}

\item{dfin.adj.up}{Position adjustment along the Y axis (dorso-ventral axis) for the dorsal fin, in pixels.}

\item{dfin.onset}{Position adjustment along the Z axis ( cranio-caudal axis) for the dorsal fin, in fraction, usually between 0 and 1.}

\item{dfin.roll}{Rotation angle of the dorsal fin around Z axis ( cranio-caudal axis). 0 for vertical, pi/2 for horizontal fins.}

\item{dfin.pitch}{Rotation angle of the dorsal fin around the X axis (bilateral axis).}

\item{dfin.yaw}{Yaw angle of dorsal fin around the Y axis (dorso-ventral axis).}

\item{dfin.thick}{Maximum thickness of the dorsal fin relative to the chord in percentage.}

\item{d2fin.adj.up}{Position adjustment along the Y (dorso-ventral axis) axis for the second dorsal fin, in pixels.}

\item{d2fin.onset}{Position adjustment along the Z axis ( cranio-caudal axis) for the secod dorsal fin, in fraction, usually between 0 and 1.}

\item{d2fin.roll}{Rotation angle of the second dorsal fin around the Z axis  (cranio-caudal axis).}

\item{d2fin.pitch}{Rotation angle of the second dorsal fin around the X axis (bilateral axis).}

\item{d2fin.thick}{Maximum thickness of the second dorsal fin relative to the chord in percentage.}

\item{ffin.adj.lat}{Position adjustment along the X axis (bilateral axis) for the forefin, in pixels.}

\item{ffin.adj.up}{Position adjustment along the Y axis (dorso-ventral axis) for the forefin, in pixels.}

\item{ffin.onset}{Position adjustment along the Z axis ( cranio-caudal axis) for the forefin, in fraction, usually between 0 and 1.}

\item{ffin.roll}{Rotation angle of the forefin around the Z axis ( cranio-caudal axis).}

\item{ffin.pitch}{Rotation angle of the forefin around the X axis (bilateral axis).}

\item{ffin.thick}{Maximum thickness of the forefin relative to the chord in percentage.}

\item{ffin.yaw}{Yaw angle of forefin around the Y axis (dorso-ventral axis).}

\item{hfin.adj.lat}{Position adjustment along the X axis (bilateral axis) for the hindfin, in pixels.}

\item{hfin.adj.up}{Position adjustment along the Y axis (dorso-ventral axis) for the hindfin, in pixels.}

\item{hfin.onset}{Position adjustment along the Z axis ( cranio-caudal axis) for the hindfin, in fraction, usually between 0 and 1.}

\item{hfin.roll}{Rotation angle of the hindfin around the Z axis ( cranio-caudal axis).}

\item{hfin.pitch}{Rotation angle of the hindfin around the X axis (bilateral axis).}

\item{hfin.thick}{Maximum thickness of the hindfin relative to the chord in percentage.}

\item{hfin.yaw}{Yaw angle of hindfin around the Y axis (dorso-ventral axis).}

\item{afin.adj.up}{Position adjustment along the Y axis (dorso-ventral axis) for the anal fin, in pixels.}

\item{afin.onset}{Position adjustment along the Z axis (cranio-caudal axis) for the anal fin, in fraction, usually between 0 and 1.}

\item{afin.roll}{Rotation angle of the anal fin around Z axis (cranio-caudal axis). 0 for vertical, pi/2 for horizontal fins.}

\item{afin.pitch}{Rotation angle of the anal fin around the X axis (bilateral axis).}

\item{afin.thick}{Maximum thickness of the anal fin relative to the chord in percentage.}

\item{afin.yaw}{Yaw angle of anal fin around the Y axis (dorso-ventral axis).}

\item{ceph.adj.up}{Position adjustment along the Y axis (dorso#-ventral axis) for the cephalofoil, in pixels.}

\item{ceph.onset}{Position adjustment along the Z axis (cranio-caudal axis) for the cephalofoil, in fraction, usually between 0 and 1.}

\item{ceph.roll}{Rotation angle of the cephalofoil around Z axis (cranio-caudal axis). pi/2 for horizontal cephalfoil.}

\item{ceph.pitch}{Rotation angle of the cephalofoil around the X axis (bilateral axis).}

\item{ceph.thick}{Maximum thickness of the cephalofoil relative to the chord in percentage.}

\item{Test.Run}{Whether to set the positions and rotations of fins for a test run}
}
\value{
A matrix containing measurement results.
}
\description{
Main function to build two 3D models from silhouettes of a marine vertebrate in a trial to bracket the body volume and surface area between the two.
}
\examples{
# Ichthyosaur Example
# Save the silhouette images for Stenopterygius in the current
# directory and make 3D models using them.
save_Stenopterygius()
paleomass()
#
# Plesiosaur Example
# Save the silhouette images for Plesiopteryx and make 3D models using them
save_Plesiopteryx()
paleomass(Folder = "Plesiopteryx", n1 = 2, n2 = 2.5, body.axis.l = 2.94,
    ffin.onset = 0.45, ffin.adj.up = 50, ffin.adj.lat = -100, ffin.pitch = pi/6, ffin.roll = pi/3,
    hfin.onset = 0.65, hfin.adj.up = 150, hfin.adj.lat = -200, hfin.pitch =  pi/7,  hfin.roll = pi/4)
#
# Hammerhead Example
# Save the silhouette images for Sphyrna and make 3D models using them
save_Sphryna()
paleomass(Folder = "Sphyrna", n1 = 1.8, n2 = 2, nn.f = 0.1, body.axis.l = 0.565,
    cfin.onset = 0.78, cfin.adj.up = 200, cfin.adj.lat = 0,
    cfin.pitch = -pi/4, cfin.thick = 20,
    ffin.onset = 0.232, ffin.adj.up = 80, ffin.adj.lat = -120, ffin.thick = 20,
    hfin.onset = 0.494, hfin.adj.up = 150, hfin.adj.lat = -100, hfin.thick = 20,
    hfin.thick=10,hfin.pitch=pi*10/180,
    dfin.onset = 0.274, dfin.adj.up = 0, dfin.thick = 20,
    d2fin.onset = 0.626, d2fin.adj.up = 75, d2fin.thick = 10,
    afin.onset = 0.603, afin.adj.up = 110, afin.thick = 10,
    ceph.roll = pi/2, ceph.adj.up = 150)
}
