% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paleomass.R
\name{paleomass}
\alias{paleomass}
\title{Main function to build two 3D models from silhouettes of a marine vertebrate in a trial to bracket the body volume and surface area between the two.}
\usage{
paleomass(
  Folder = "Stenopterygius",
  File.Ext = c(".png", ".jpg", ".jpeg", ".bmp", ""),
  BodyL.File = "BodyL",
  BodyV.File = "BodyV",
  Cfin.File = "Cfin",
  Dfin.File = "Dfin",
  D2fin.File = "D2fin",
  Ffin.File = "Ffin",
  Hfin.File = "Hfin",
  Afin.File = "Afin",
  Ceph.File = "Ceph",
  n1 = 2,
  n2 = 3,
  fork.l = 1.19,
  dens = 0.95,
  Save.Csv = T,
  Interpolate = T,
  nn.b = 0.1,
  nn.f = 0.6,
  Use.Weight = F,
  Add.Tip = T,
  tip.radius = 1e-04,
  a.per.slice = 181,
  Plot.Result.Mesh = T,
  Save.Part.Mesh = F,
  Save.Total.Mesh = T,
  N.Mesh.Window = c("two", "one"),
  Mesh.Col = "royalblue",
  Body.Y.Centering = T,
  N.Body.Slice = c("round", "asis", "provided"),
  n.body.slice = 1000,
  N.Fin.Slice = c("asis", "round", "provided"),
  n.fin.slice = 500,
  cfin.adj.up = -5,
  cfin.onset = 870,
  cfin.rot = 0,
  cfin.sweep = 0,
  cfin.thick = 20,
  dfin.adj.up = 175,
  dfin.onset = 440,
  dfin.sweep = 0,
  dfin.thick = 20,
  d2fin.adj.up = 165,
  d2fin.onset = 600,
  d2fin.sweep = 0,
  d2fin.thick = 10,
  ffin.spread = pi/4,
  ffin.sweep = pi/6,
  ffin.adj.med = 30,
  ffin.adj.up = -67,
  ffin.onset = 300,
  ffin.thick = 20,
  hfin.spread = pi/6,
  hfin.sweep = pi/7,
  hfin.adj.med = 30,
  hfin.adj.up = -27,
  hfin.onset = 600,
  hfin.thick = 20,
  afin.adj.up = -25,
  afin.onset = 650,
  afin.sweep = 0,
  afin.thick = 10,
  ceph.adj.up = -5,
  ceph.onset = 870,
  ceph.rot = 0,
  ceph.sweep = 0,
  ceph.thick = 30,
  Test.Run = F
)
}
\value{
A matrix containing measurement results.
}
\description{
Main function to build two 3D models from silhouettes of a marine vertebrate in a trial to bracket the body volume and surface area between the two.
}
\examples{
# Ichthyosaur Example
# Save the silhouette images for Stenopterygius in the current
# directory and make 3D models using them.
save_Stenopterygius()
paleomass()
#
# Plesiosaur Example
# Save the silhouette images for Plesiosaurus and make 3D models using them
save_Plesiosaurus()
paleomass(Folder="Plesiosaurus",fork.l=2.94,ffin.onset=450,
    hfin.onset=650,ffin.adj.med=10,hfin.adj.med=40,ffin.adj.up=-160,
    hfin.adj.up=-160,ffin.spread=pi/3,hfin.spread=pi/4)
#
# Hammerhead Example
# Save the silhouette images for Sphyrna and make 3D models using them
save_Sphryna()
paleomass(Folder="Sphyrna",fork.l=2,n2=1.8,
    ffin.onset=300, ffin.adj.up=-120,ffin.adj.med=40,ffin.spread=pi/4,
    ffin.sweep=-pi*0/180,
    hfin.onset=550,hfin.adj.up=10,hfin.adj.med=40,hfin.spread=pi/4,
    hfin.thick=10,hfin.sweep=pi*5/180,
    cfin.onset=920,cfin.adj.up=35,
    dfin.onset=350,dfin.adj.up=190,dfin.sweep=pi*5/180,
    d2fin.onset=800,d2fin.adj.up=200,,d2fin.sweep=pi*5/180,d2fin.thick=10,
    afin.onset=800,afin.adj.up=60,afin.sweep=pi*20/180,afin.thick=10,
    ceph.onset=-45,ceph.adj.up=120, ceph.rot=pi/2)
}
